refactor(state_model): extract generic JSON/JSONL I/O into json_io.rs
persistence.rs was 514 lines — the third and last file the architecture
audit flagged as pressing the 600-line ceiling. Most of its bulk was
not schema-aware logic, but ~140 lines of type-parametric filesystem
helpers (ensure_json_file / ensure_text_file / write_json_file /
append_json_line / read_json_file / read_json_lines / write_atomic)
that know nothing about V3 semantics. Splitting by concern — generic
I/O vs schema orchestration — is the cleanest cut available.

Move those seven helpers to src/state_model/json_io.rs behind
`pub(super)` visibility. persistence.rs retains the schema-aware
surface (V3StatePaths constructors, ensure_layout, load_or_init_snapshot,
load_snapshot, the per-file load/save pairs, validate_schema_version,
tests) and now imports the generic primitives.

Result:
  persistence.rs  514 → 403 L  (−22%, well under the ceiling)
  json_io.rs      new → 145 L  (single-concern module)

Scope decisions:
- Kept `validate_schema_version` in persistence.rs — it depends on
  V3_SCHEMA_VERSION and only makes sense inside the V3 context, so
  it's not a generic helper.
- Visibility chose `pub(super)` rather than `pub(crate)` — only the
  state_model tree should reach these helpers; nothing outside v3
  persistence has any business calling them.

Known follow-up (intentionally out of scope):
- A near-identical `write_atomic` still lives in src/state/mod.rs for
  pre-v3 pending-ops journaling. It predates this refactor and
  deduping it would mean changing the call sites in state/ops.rs
  too. Tracked separately — single-batch scope discipline.

Tested: make fmt-check, make lint (clippy -D warnings), make test
(75/75 pass; the two persistence.rs unit tests
`builds_expected_v3_paths` and `query_helpers_link_bindings_targets_and_projections`
still pass, confirming the helper move didn't alter V3 disk layout).
Scope-risk: narrow (3 files, visibility strictly pub(super)).
Reversibility: clean — inline json_io.rs back into persistence.rs.
Confidence: high.
